requirements.txt
# ── Summarize AI — Dependencies ──────────────────────────────
# Install: pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────

# Core
flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.2.2

# AI / LLM
groq==0.37.1
langchain==1.2.15
langchain-core==1.2.28
langchain-groq==1.1.2
langgraph==1.1.6
langgraph-checkpoint==4.0.1

# Data Validation
pydantic==2.12.5

# Utilities
requests==2.33.1
tiktoken==0.12.0
